requirements.txt
asgiref==3.9.1
cffi==1.17.1
crispy-bootstrap4==2025.6
cryptography==45.0.5
Django==5.2.4
django-crispy-forms==2.4
mysql==0.0.3
mysqlclient==2.2.7
pillow==11.3.0
pycparser==2.22
PyMySQL==1.1.1
sqlparse==0.5.3
tzdata==2025.2
